#5 Move common items to utils module
