+++ xarray_units/utils.py
@@ -4,7 +4,27 @@ __all__ = [
     "UnitsExistError",
     "UnitsNotFoundError",
     "UnitsNotValidError",
+    "units_of",
 ]
+
+
+# standard library
+from typing import Any, Optional, TypeVar, Union, overload
+
+
+# dependencies
+from astropy.units import Equivalency, Quantity, Unit, UnitBase
+from xarray import DataArray
+
+
+# type hints
+TDataArray = TypeVar("TDataArray", bound=DataArray)
+Equivalencies = Optional[Equivalency]
+UnitsLike = Union[UnitBase, str]
+
+
+# constants
+UNITS_ATTR = "units"
 
 
 class UnitsError(Exception):
@@ -35,3 +55,31 @@ class UnitsNotValidError(UnitsError):
     """Units are not valid for a DataArray."""
 
     pass
+
+
+@overload
+def units_of(obj: Quantity) -> UnitBase:
+    ...
+
+
+@overload
+def units_of(obj: Any) -> Optional[UnitBase]:
+    ...
+
+
+def units_of(obj: Any) -> Optional[UnitBase]:
+    """Return units of an object if they exist and are valid."""
+    if isinstance(obj, Quantity):
+        if isinstance(units := obj.unit, UnitBase):
+            return units
+
+        raise UnitsNotValidError(repr(obj))
+
+    if isinstance(obj, DataArray):
+        if (units := obj.attrs.get(UNITS_ATTR)) is None:
+            return None
+
+        if isinstance(units := Unit(units), UnitBase):  # type: ignore
+            return units
+
+        raise UnitsNotValidError(repr(obj))
